Remove more unused files and rename saveData to utils
